+++ pkg/machinepool/validate_create_test.go
@@ -10,11 +10,10 @@ import (
 	"github.com/giantswarm/apiextensions/v3/pkg/label"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	capzexp "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 	capi "sigs.k8s.io/cluster-api/api/v1alpha3"
+	capiexp "sigs.k8s.io/cluster-api/exp/api/v1alpha3"
 
 	builder "github.com/giantswarm/azure-admission-controller/internal/test/machinepool"
 	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
@@ -30,7 +29,7 @@ const (
 func TestMachinePoolCreateValidate(t *testing.T) {
 	type testCase struct {
 		name         string
-		machinePool  []byte
+		machinePool  *capiexp.MachinePool
 		vmType       string
 		errorMatcher func(err error) bool
 	}
@@ -38,43 +37,43 @@ func TestMachinePoolCreateValidate(t *testing.T) {
 	testCases := []testCase{
 		{
 			name:         "case 0: instance type supporting [1,2,3], requested [1]",
-			machinePool:  builder.BuildMachinePoolAsJson(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"1"})),
+			machinePool:  builder.BuildMachinePool(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"1"})),
 			vmType:       "Standard_A2_v2",
 			errorMatcher: nil,
 		},
 		{
 			name:         "case 1: instance type supporting [1,2], requested [3]",
-			machinePool:  builder.BuildMachinePoolAsJson(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"3"})),
+			machinePool:  builder.BuildMachinePool(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"3"})),
 			vmType:       "Standard_A4_v2",
 			errorMatcher: IsUnsupportedFailureDomainError,
 		},
 		{
 			name:         "case 2: instance type supporting [1,2], requested [2,3]",
-			machinePool:  builder.BuildMachinePoolAsJson(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"2,3"})),
+			machinePool:  builder.BuildMachinePool(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"2,3"})),
 			vmType:       "Standard_A4_v2",
 			errorMatcher: IsUnsupportedFailureDomainError,
 		},
 		{
 			name:         "case 3: instance type supporting [], requested [1]",
-			machinePool:  builder.BuildMachinePoolAsJson(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"1"})),
+			machinePool:  builder.BuildMachinePool(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{"1"})),
 			vmType:       "Standard_A8_v2",
 			errorMatcher: IsUnsupportedFailureDomainError,
 		},
 		{
 			name:         "case 4: instance type supporting [], requested []",
-			machinePool:  builder.BuildMachinePoolAsJson(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{})),
+			machinePool:  builder.BuildMachinePool(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{})),
 			vmType:       "Standard_A8_v2",
 			errorMatcher: nil,
 		},
 		{
 			name:         "case 5: AzureMachinePool does not exist",
-			machinePool:  builder.BuildMachinePoolAsJson(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{})),
+			machinePool:  builder.BuildMachinePool(builder.AzureMachinePool(machinePoolName), builder.FailureDomains([]string{})),
 			vmType:       "",
 			errorMatcher: IsAzureMachinePoolNotFound,
 		},
 		{
 			name:         "case 6: Wrong Organization",
-			machinePool:  builder.BuildMachinePoolAsJson(builder.AzureMachinePool(machinePoolName), builder.Organization("wrongorg")),
+			machinePool:  builder.BuildMachinePool(builder.AzureMachinePool(machinePoolName), builder.Organization("wrongorg")),
 			vmType:       "",
 			errorMatcher: generic.IsNodepoolOrgDoesNotMatchClusterOrg,
 		},
@@ -223,8 +222,9 @@ func TestMachinePoolCreateValidate(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			admit, err := NewCreateValidator(CreateValidatorConfig{
+			handler, err := NewWebhookHandler(WebhookHandlerConfig{
 				CtrlClient: ctrlClient,
+				Decoder:    unittest.NewFakeDecoder(),
 				Logger:     newLogger,
 				VMcaps:     vmcaps,
 			})
@@ -232,8 +232,8 @@ func TestMachinePoolCreateValidate(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			// Run admission request to validate AzureConfig updates.
-			err = admit.Validate(ctx, getCreateAdmissionRequest(tc.machinePool))
+			// Run validating webhook handler on MachinePool creation.
+			err = handler.OnCreateValidate(ctx, tc.machinePool)
 
 			// Check if the error is the expected one.
 			switch {
@@ -258,22 +258,6 @@ func NewStubAPI(stubbedSKUs map[string]compute.ResourceSku) vmcapabilities.API {
 	return &StubAPI{stubbedSKUs: stubbedSKUs}
 }
 
-func (s *StubAPI) List(ctx context.Context, filter string) (map[string]compute.ResourceSku, error) {
+func (s *StubAPI) List(_ context.Context, _ string) (map[string]compute.ResourceSku, error) {
 	return s.stubbedSKUs, nil
-}
-
-func getCreateAdmissionRequest(newMP []byte) *v1beta1.AdmissionRequest {
-	req := &v1beta1.AdmissionRequest{
-		Resource: metav1.GroupVersionResource{
-			Version:  "exp.infrastructure.cluster.x-k8s.io/v1alpha3",
-			Resource: "azuremachinepool",
-		},
-		Operation: v1beta1.Create,
-		Object: runtime.RawExtension{
-			Raw:    newMP,
-			Object: nil,
-		},
-	}
-
-	return req
 }