+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `HttpHandlerFactory` for creating HTTP handlers that are using new webhook handlers.
 - Cluster webhook handler that replaces mutators and validators.
 - AzureCluster webhook handler that replaces mutators and validators.
+- MachinePool webhook handler that replaces mutators and validators.
 
 ### Changed